server/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
gspread==6.1.0
google-auth==2.29.0
python-dotenv==1.0.1
google-generativeai==0.4.1
pydantic==2.6.4
python-multipart==0.0.9
